refactor: 프로젝트 구조 변경
- src 디렉토리 생성하고 components, libs, styles 디렉토리 이동
